Rename UnityYamlDisableStrategy
